+++ bundle/src/test/java/com/adobe/acs/commons/ondeploy/scripts/OnDeployScriptTestExampleSuccess1.java
@@ -1,0 +1,10 @@
+package com.adobe.acs.commons.ondeploy.scripts;
+
+import com.adobe.acs.commons.ondeploy.scripts.OnDeployScriptBase;
+
+public class OnDeployScriptTestExampleSuccess1 extends OnDeployScriptBase {
+    @Override
+    protected void execute() throws Exception {
+        logger.info("Executing test script: OnDeployScriptTestExampleSuccess1");
+    }
+}